backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==1.0.0
mistralai==0.0.7
pytest==7.4.0
httpx==0.24.1